Implement type casting and confirm array length in using_as.rs

+++ exercises/23_conversions/using_as.rs
@@ -4,8 +4,8 @@
 
 fn average(values: &[f64]) -> f64 {
     let total = values.iter().sum::<f64>();
-    // TODO: Make a conversion before dividing.
-    total / values.len()
+    // Make a conversion before dividing: cast the length to f64 to avoid integer division.
+    total / (values.len() as f64)
 }
 
 fn main() {
